- (agent/mibgroup/dummy/dummy.c agent/mibgroup/examples/example.c  agent/mibgroup/examples/example.h  agent/mibgroup/examples/wombat.c agent/mibgroup/misc/ipfwacc.c  agent/mibgroup/smux/smux.c  agent/mibgroup/snmpv3/snmpMPDStats.c  agent/mibgroup/snmpv3/usmStats.c  agent/mibgroup/snmpv3/usmUser.c snmplib/snmp_auth.c) more size_t fixes
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2616 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/dummy/dummy.c
@@ -109,9 +109,9 @@ void dummy_free_config (void) {
 int
 header_dummy(struct variable *vp,
 	     oid *name,
-	     int *length,
+	     size_t *length,
 	     int exact,
-	     int *var_len,
+	     size_t *var_len,
 	     WriteMethod **write_method) 
 {
 #define DUMMY_NAME_LENGTH	8
@@ -146,9 +146,9 @@ header_dummy(struct variable *vp,
 u_char	*
 var_dummy(struct variable *vp,
 	  oid *name,
-	  int *length,
+	  size_t *length,
 	  int exact,
-	  int *var_len,
+	  size_t *var_len,
 	  WriteMethod **write_method) 
 {
